In a move to rather focus more on this 'project' labelled : "Java_Data_Structures_and_Algorithms_Masterclass" and rather zooming out of the "DataStructuresRevision",
For starters we are bringing in "List" : AbstractSequentialList, ArrayList, CopyOnWriteArrayList, LinkedList and Vector.

As a baseline ; there is also the Utilities package that has been added. Basically the Utilities package is there to help with refactoring all the methods in one place and improve the 'DRY'-ness of the code. DRY is Dont Repeat Yourself.